+++ electron/src/electron-utils/utilsSQLite.ts
@@ -192,7 +192,7 @@ export class UtilsSQLite {
    */
   public async beginTransaction(db: any, isOpen: boolean): Promise<void> {
     // eslint-disable-next-line no-async-promise-executor
-    return new Promise( (resolve, reject) => {
+    return new Promise((resolve, reject) => {
       const msg = 'BeginTransaction: ';
       if (!isOpen) {
         return Promise.reject(`${msg}database not opened`);
@@ -204,7 +204,7 @@ export class UtilsSQLite {
           reject(`${msg}${err.message}`);
         }
         resolve();
-      }); 
+      });
     });
   }
   /**
@@ -831,7 +831,7 @@ export class UtilsSQLite {
   }
   public async getJournalMode(mDB: any): Promise<string> {
     let resQuery: any[] = [];
-    let retMode = "delete";
+    let retMode = 'delete';
     const query = `PRAGMA journal_mode;`;
     try {
       resQuery = await this.queryAll(mDB, query, []);
@@ -844,8 +844,6 @@ export class UtilsSQLite {
     } catch (err) {
       return Promise.reject('GetJournalMode: ' + `${err}`);
     }
-
-
   }
   /**
    * GetTableColumnNamesTypes